+++ app/src/components/pages/collectionPage/collectionPage.tsx
@@ -13,12 +13,22 @@ import {
   Icon,
   Pagination,
 } from "@nypl/design-system-react-components";
-import React from "react";
-import Filters from "../../search/filters";
+import React, { useRef } from "react";
+import Filters from "../../search/filters/filters";
 import { headerBreakpoints } from "@/src/utils/breakpoints";
-import { CardsGrid } from "../../grids/cardsGrid";
-import { displayResults } from "@/src/utils/utils";
-import { CARDS_PER_PAGE } from "@/src/config/constants";
+import { displayResults, totalNumPages } from "@/src/utils/utils";
+import {
+  CARDS_PER_PAGE,
+  DEFAULT_PAGE_NUM,
+  DEFAULT_SEARCH_SORT,
+  DEFAULT_SEARCH_TERM,
+  SEARCH_SORT_LABELS,
+} from "@/src/config/constants";
+import SearchCardsGrid from "../../grids/searchCardsGrid";
+import { GeneralSearchManager } from "@/src/utils/searchManager";
+import { stringToFilter } from "@/src/context/SearchProvider";
+import { usePathname, useRouter } from "next/navigation";
+import SortMenu from "../../sortMenu/sortMenu";
 
 const textLink = (href, text) => {
   return (
@@ -37,7 +47,23 @@ const textLink = (href, text) => {
   );
 };
 
-const CollectionPage = ({ slug, data }) => {
+const CollectionPage = ({ slug, data, searchParams }) => {
+  const headingRef = useRef<HTMLHeadingElement>(null);
+  const searchManager = new GeneralSearchManager({
+    initialPage: Number(searchParams?.page) || DEFAULT_PAGE_NUM,
+    initialSort: searchParams?.sort || DEFAULT_SEARCH_SORT,
+    initialFilters: stringToFilter(searchParams?.filters),
+    initialKeywords: searchParams?.keywords || DEFAULT_SEARCH_TERM,
+  });
+  const totalPages = totalNumPages(data.numResults, CARDS_PER_PAGE);
+  const { push } = useRouter();
+  const pathname = usePathname();
+
+  const updateURL = async (queryString) => {
+    push(`${pathname}?${queryString}`);
+    headingRef.current?.focus();
+  };
+
   return (
     <Box id="mainContent">
       <Box
@@ -62,7 +88,7 @@ const CollectionPage = ({ slug, data }) => {
           >
             {slug}
           </Heading>
-          <Filters headingText="Search this collection" />
+          <Filters headingText="Search this collection" ref={headingRef} />
         </Box>
       </Box>
       <Box
@@ -73,7 +99,6 @@ const CollectionPage = ({ slug, data }) => {
           paddingRight: { base: "m", xl: "s" },
         }}
       >
-        <HorizontalRule />
         <Flex alignContent="center" alignItems="center" gap="xs">
           <Text size="subtitle2" sx={{ margin: 0, fontWeight: 400 }}>
             Filters applied:
@@ -206,13 +231,30 @@ const CollectionPage = ({ slug, data }) => {
                 }}
               />
             </Flex>
-            <Heading
+            <Flex
               marginTop="xl"
-              size="heading5"
-              tabIndex={-1}
-            >{`Displaying ${displayResults(data.numResults, CARDS_PER_PAGE, 1)}
+              marginBottom="s"
+              alignItems="center"
+              justifyContent="space-between"
+            >
+              <Heading
+                size="heading5"
+                ref={headingRef}
+                tabIndex={-1}
+                margin="0"
+              >{`Displaying ${displayResults(
+                data.numResults,
+                CARDS_PER_PAGE,
+                1
+              )}
                                 results`}</Heading>
-            <CardsGrid records={data} />
+              <SortMenu
+                options={SEARCH_SORT_LABELS}
+                searchManager={searchManager}
+                updateURL={updateURL}
+              />
+            </Flex>
+            <SearchCardsGrid keywords={data.keyword} results={data.results} />
             <Flex marginTop="xxl" marginBottom="xxl" alignContent="center">
               <Link
                 minWidth="100px"